+++ src/sll/include/sll/search_path.h
@@ -27,25 +27,61 @@
 
 
 /**
- * \flags func
- * \name sll_create_search_path
- * \group env
+ * \flags macro var
+ * \name SLL_SEARCH_PATH_FLAG_BEFORE
+ * \group search-path
  * \desc Docs!
- * \arg const sll_string_t* src
- * \arg sll_search_path_t* o
+ * \type sll_flags_t
  */
-__SLL_EXTERNAL void sll_create_search_path(const sll_string_t* src,sll_search_path_t* o);
+#define SLL_SEARCH_PATH_FLAG_BEFORE 1
+
+
+/**
+ * \flags macro var
+ * \name SLL_SEARCH_PATH_FLAG_AFTER
+ * \group search-path
+ * \desc Docs!
+ * \type sll_flags_t
+ */
+#define SLL_SEARCH_PATH_FLAG_AFTER 2
 
 
 
 /**
  * \flags func
  * \name sll_free_search_path
- * \group env
+ * \group search-path
  * \desc Docs!
  * \arg sll_search_path_t* sp
  */
 __SLL_EXTERNAL void sll_free_search_path(sll_search_path_t* sp);
+
+
+
+/**
+ * \flags func
+ * \name sll_search_path_create
+ * \group search-path
+ * \desc Docs!
+ * \arg const sll_string_t* src
+ * \arg sll_search_path_t* o
+ */
+__SLL_EXTERNAL void sll_search_path_create(const sll_string_t* src,sll_search_path_t* o);
+
+
+
+/**
+ * \flags check_output func
+ * \name sll_search_path_create
+ * \group search-path
+ * \desc Docs!
+ * \arg const sll_search_path_t* sp
+ * \arg const sll_string_t* nm
+ * \arg sll_flags_t fl
+ * \arg sll_string_t* o
+ * \ret sll_bool_t
+ */
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_bool_t sll_search_path_find(const sll_search_path_t* sp,const sll_string_t* nm,sll_flags_t fl,sll_string_t* o);