+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - CLI option to load an audit library: `-L` or `--audit-library`
 - Implemented `SLL_ABI_AUDIT_CALL`, `SLL_ABI_AUDIT_DEINIT`, `SLL_ABI_AUDIT_INIT`, `SLL_ABI_DEINIT`, `SLL_ABI_INIT` and `SLL_ABI_NAME`
-- Implemented `sll_search_path_t`, `SLL_SEARCH_PATH_SPLIT_CHAR`, `sll_env_path`, `sll_create_search_path` and `sll_free_search_path`
+- Implemented `sll_search_path_t`, `SLL_SEARCH_PATH_SPLIT_CHAR`, `SLL_SEARCH_PATH_FLAG_BEFORE`, `SLL_SEARCH_PATH_FLAG_AFTER`, `sll_env_path`, `sll_free_search_path`, `sll_search_path_create` and `sll_search_path_find`
 - Implemented `statistics$correlation` ([#260])
 - Object creation code to build an array of string pointers: `x`
 
@@ -15,6 +15,11 @@
 | Name | Format | Arguments |
 |--|--|--|
 | `sll.cli.start` | `a` | `arguments` |
+
+### Changed
+
+- Fixed empty string arguments in `sll_api_path_join`
+- Fixed executable path expansion in `sll_api_process_start`
 
 ## [0.7.11] - 2022-04-02